Refactor examples; automatically build enums
