+++ rust-package/brainflow/Cargo.toml
@@ -26,6 +26,11 @@ serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.68"
 thiserror = "1.0.29"
 
+[dev-dependencies]
+approx = "0.5.0"
+
 [build-dependencies]
 bindgen = { version = "0.59.1", optional = true }
+convert_case = "0.4.0"
 fs_extra = "1.2.0"
+regex = "1.5.4"
