+++ rust-package/brainflow/examples/markers.rs
@@ -5,8 +5,7 @@ use brainflow::board_shim::BoardShim;
 use brainflow::brainflow_input_params::BrainFlowInputParamsBuilder;
 use brainflow::BoardIds;
 
-#[test]
-fn markers() {
+fn main() {
     brainflow::board_shim::enable_dev_board_logger().unwrap();
     let params = BrainFlowInputParamsBuilder::default().build();
     let board = BoardShim::new(BoardIds::SyntheticBoard as i32, params).unwrap();
@@ -22,9 +21,4 @@ fn markers() {
 
     assert_eq!(data.len(), 32);
     println!("{:?}", data);
-
-    // check that the first array in data are sequential numbers
-    for (d, d_plus_1) in data[0].iter().zip(data[0].iter().skip(1)) {
-        assert_eq!(d + 1.0, *d_plus_1);
-    }
 }
