--- rust-package/brainflow/src/agg_operations.rs
@@ -1,6 +1,0 @@
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, FromPrimitive)]
-pub enum AggOperations {
-    Mean = 0,
-    Median = 1,
-    Each = 2,
-}
