--- rust-package/brainflow/src/brainflow_classifiers.rs
@@ -1,7 +1,0 @@
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, FromPrimitive)]
-pub enum BrainFlowClassifiers {
-    Regression = 0,
-    Knn = 1,
-    Svm = 2,
-    Lda = 3,
-}
