--- rust-package/brainflow/src/brainflow_exit_codes.rs
@@ -1,27 +1,0 @@
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, FromPrimitive)]
-pub enum BrainFlowExitCodes {
-    StatusOk = 0,
-    PortAlreadyOpenError = 1,
-    UnableToOpenPortError = 2,
-    SetPortError = 3,
-    BoardWriteError = 4,
-    IncommingMsgError = 5,
-    InitialMsgError = 6,
-    BoardNotReadyError = 7,
-    StreamAlreadyRunError = 8,
-    InvalidBufferSizeError = 9,
-    StreamThreadError = 10,
-    StreamThreadIsNotRunning = 11,
-    EmptyBufferError = 12,
-    InvalidArgumentsError = 13,
-    UnsupportedBoardError = 14,
-    BoardNotCreatedError = 15,
-    AnotherBoardIsCreatedError = 16,
-    GeneralError = 17,
-    SyncTimeoutError = 18,
-    JsonNotFoundError = 19,
-    NoSuchDataInJsonError = 20,
-    ClassifierIsNotPreparedError = 21,
-    AnotherClassifierIsPreparedError = 22,
-    UnsupportedClassifierAndMetricCombinationError = 23,
-}
