--- rust-package/brainflow/src/detrend_operations.rs
@@ -1,6 +1,0 @@
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, FromPrimitive)]
-pub enum DetrendOperations {
-    None = 0,
-    Constant = 1,
-    Linear = 2,
-}
