+++ rust-package/brainflow/src/ffi.rs
@@ -1,4 +1,5 @@
 #![allow(dead_code)]
 pub mod board_controller;
+pub mod constants;
 pub mod data_handler;
 pub mod ml_model;
