+++ rust-package/brainflow/src/ffi/board_controller.rs
@@ -2,7 +2,6 @@
 
 #![allow(non_camel_case_types)]
 
-
 extern "C" {
     pub fn get_board_descr(
         board_id: ::std::os::raw::c_int,
