+++ rust-package/brainflow/src/ffi/data_handler.rs
@@ -2,7 +2,6 @@
 
 #![allow(non_camel_case_types)]
 
-
 extern "C" {
     pub fn perform_lowpass(
         data: *mut f64,
