--- rust-package/brainflow/src/filter_types.rs
@@ -1,6 +1,0 @@
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, FromPrimitive)]
-pub enum FilterTypes {
-    Butterworth = 0,
-    ChebyshevType1 = 1,
-    Bessel = 2,
-}
