--- rust-package/brainflow/src/ip_protocol_type.rs
@@ -1,6 +1,0 @@
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, FromPrimitive)]
-pub enum IpProtocolType {
-    None = 0,
-    Udp = 1,
-    Tcp = 2,
-}
