+++ rust-package/brainflow/src/lib.rs
@@ -1,55 +1,49 @@
+#[cfg(test)]
+#[macro_use]
+extern crate approx;
+
 extern crate num;
 #[macro_use]
 extern crate num_derive;
 
-use error::{Error};
+use error::Error;
 
-pub mod board_ids;
-pub mod agg_operations;
-pub mod brainflow_classifiers;
-pub mod detrend_operations;
-pub mod error;
-pub mod filter_types;
-pub mod ip_protocol_type;
-pub mod log_levels;
-pub mod noise_types;
-pub mod window_functions;
-pub mod brainflow_exit_codes;
 /// The primary interface to all boards.
 pub mod board_shim;
 /// Input parameters for [board_shim::BoardShim].
 pub mod brainflow_input_params;
+pub mod error;
 
+/// Input parameters for [ml_model::MLModel].
+pub mod brainflow_model_params;
 /// Methods for signal processig.
 pub mod data_filter;
 mod ffi;
-/// Input parameters for [ml_model::MLModel].
-pub mod brainflow_model_params;
 /// Used to calculate derivative metrics from raw data.
 pub mod ml_model;
 
-/// Enum to store all supported Board Ids.
-pub use board_ids::BoardIds;
-/// Enum to store all supported Agg Operations.
-pub use agg_operations::AggOperations;
-/// Enum to store all supported BrainFlow Classifiers.
-pub use brainflow_classifiers::BrainFlowClassifiers;
-/// Enum to store all supported Detrend Operations.
-pub use detrend_operations::DetrendOperations;
 /// Store all supported BrainFlow Errors.
 pub use error::BrainFlowError;
-/// Enum to store all supported Filter Types.
-pub use filter_types::FilterTypes;
-/// Enum to store all supported Ip Protocol Types.
-pub use ip_protocol_type::IpProtocolType;
-/// Enum to store all supported Log Levels.
-pub use log_levels::LogLevels;
-/// Enum to store all supported Noise Types.
-pub use noise_types::NoiseTypes;
-/// Enum to store all supported Window Functions.
-pub use window_functions::WindowFunctions;
+/// Enum to store all supported Agg Operations.
+pub use ffi::constants::AggOperations;
+/// Enum to store all supported Board Ids.
+pub use ffi::constants::BoardIds;
+/// Enum to store all supported BrainFlow Classifiers.
+pub use ffi::constants::BrainFlowClassifiers;
 /// Enum to store all Brainflow Exit Codes.
-pub use brainflow_exit_codes::BrainFlowExitCodes;
+pub use ffi::constants::BrainFlowExitCodes;
+/// Enum to store all supported Detrend Operations.
+pub use ffi::constants::DetrendOperations;
+/// Enum to store all supported Filter Types.
+pub use ffi::constants::FilterTypes;
+/// Enum to store all supported Ip Protocol Types.
+pub use ffi::constants::IpProtocolType;
+/// Enum to store all supported Log Levels.
+pub use ffi::constants::LogLevels;
+/// Enum to store all supported Noise Types.
+pub use ffi::constants::NoiseTypes;
+/// Enum to store all supported Window Functions.
+pub use ffi::constants::WindowFunctions;
 
 type BrainFlowExitCode = i32;
