--- rust-package/brainflow/src/noise_types.rs
@@ -1,5 +1,0 @@
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, FromPrimitive)]
-pub enum NoiseTypes {
-    Fifty = 0,
-    Sixty = 1,
-}
